requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
websockets==13.1
loguru==0.7.2
paramiko==3.4.0

# CrewAI multi-agent framework (requires Python 3.10+ — optional)
# crewai>=1.14.0
openai>=1.0.0
xmltodict>=0.13.0

# MCP client (optional)
# fastmcp>=0.4.0

# Auth service
pyotp>=2.9.0
qrcode>=7.4
PyJWT>=2.8.0
bcrypt>=4.1.0
Pillow>=10.0.0

# Tests
pytest>=7.0.0
pytest-asyncio>=0.23.0
